devices: serial: Support UnixStream for data source and destination
Add 'type=unix-stream' in a serial device flag to use a UnixStream socket
for console output. This option can be used with a new flag
'input-unix-stream=true' so the socket will be used for both input and
output.

'type=unix-stream' is similar to the existing option 'type=unix' but the one uses
Unix datagram and is only for output.

To use this feature, open a terminal and run socat first:
$ socat UNIX-LISTEN:/tmp/console.sock STDIO
Then, start crosvm in another terminal with this flag:
--serial=hardware=virtio-console,num=1,type=unix-stream,path=/tmp/console.sock,input-unix-stream=true,console=true

BUG=b:380397975
TEST=run virtio-console with type=unix-stream,input-unix-stream=true

Change-Id: I500b249781c0679d090a0dcbbd0c594c54efb4d2
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/6040758
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Commit-Queue: Keiichi Watanabe <keiichiw@chromium.org>