Autoupdate of 2 submodules
*external/double-conversion updated to 3bfbd1930ef17e76e1afc9de87f0c5463c0fdec8
Allow more packages to depend on libdoubleconversion

This is a preliminary step required for the ongoing
libchrome uprev in Android. More recent versions of
libchrome depend on libdoubleconversion instead of
shipping their own conversion functions.

For this reason, we need to allow packages that depend on
libchrome to also depend on libdoubleconversion.

Test: m libdoubleconversion
Bug: 360917504

Change-Id: I87e63c9c5029f3b3d07225e754034018a39116f8

*packages/modules/common updated to d6dd6a99ca09a3216a9d6a324f6f954c417a6932
Merge "Add libdoubleconversion to new-allowed-deps.txt" into main