Fix an untracked selinux denial
Aconfigd needs to scan all files and dirs under /apex (apex mount dir)
to find all the mainline modules. It does that by read dir, then check
if each entry is a dir, if it is a dir and then check if it has storage
files.

Due to this, it needs to get attributes of all files directly under /apex.
This caused untracked selinux denial when aconfigd tries to access apex
info file which is directly under /apex. Given that we don't really need
this permission, add the dontaudit rule.

Bug: b/381889824
Change-Id: Idb744823ede2e55d0963ad0611ef409eca6b9790