Port resultDB protos to Tradefed
As discussed, we will leave the resultDB protos in /tools/tradefederation/core for now, as TF is the only client that uses ResultDB. We will move out to a more suitable place when we have other clients.

Manually copying the protos for now, we will set up Copybara rules when we confirm things work as expected.

In this CL:
- Copy relevant resultDB protos from https://source.chromium.org/chromium/infra/infra_superproject/+/main:infra/go/src/go.chromium.org/luci/resultdb/proto/v1/
- Modify Copyright and import paths
- Create Soong rules.

Change-Id: I82d78fc96679476a819845b39d592223c0d7a0b6