Autoupdate of 2 submodules
*hardware/interfaces updated to 4991fddcf534bb44a00d4a17e7366c4f4d1ece19
Unlocked devices should have nonnormal DICE chains

Adds a test that checks two system properties for an indication that the
device is unlocked. If the DICE chain is not degenerate, then the test
checks that the mode on the leaf certificate in the DICE chain is not
normal.

Test: atest VtsHalRemotelyProvisionedComponentTargetTest
      atest libkeymint_remote_prov_support_test
Bug: 377829077
Change-Id: I65a9845f32d673833ccd2993d59f9ed899d2da11

*tools/security updated to 61dc58d00debe0f0fef24b257c37bab533e32a84
Merge "Unlocked devices should have nonnormal DICE chains" into main