Preload Cronet's shared library into Zygote [reland aosp/3416509]
After several A/B experiments runs, we found out that there's no
regression on boot-time to preload Cronet into the zygote (See
b/380349437#comment19).

This change is landing behind a build-time flag in order to carefully
roll this out as we have seen some memory regressions (b/385120248).

It's expected that this change will increase the memory consumption
of the Zygote of around (2MB rss). However, this is acceptable as this
change will drastically improve start-up time for apps performing
networking transactions using HttpEngine.

Test: Tested locally by flashing an image with preloaded shared library
and managed to send a GET request to www.google.com, verified that the
loading time has decreased.
Bug: 380349437

Change-Id: I07e0d8139fea0b7bde119b321dc8f7e8fce0bd65