tp: don't parent thread tracks to process tracks
IMO this was a mistake and never should have been done: the notion of
threads nesting under processes is purely a UI construct, doing it at
the proto/TP level is a confusion of two concepts. Otherwise,
why are processes themselves not nested under the global scope? It also
doesn't scale to other "scopes" (e.g. cpus, gpus, linux devices etc)

This should have very minimal impact to people's queries: I would argue
that relying on this was a bug in the first place,

Change-Id: Ib455c1a3f721fb5803cb3ca6c8d22bc175521fd4