Autoupdate of 3 submodules
*build/soong updated to 5dbf1beedb1f1b64faca349580115422560b8ebb
Merge "rust: Force all dependencies to be linked, used or not" into main
*external/rust/android-crates-io updated to be7611426b94dab09821b99b5c01125235bab9b2
Regenerate repo

Exact command:
./crate_tool regenerate --all --exclude arbitrary,rustc-demangle-capi

arbitrary is failing to build because cargo is fetching a version of
arbitrary-derive that is too new.

rustc-demangle-capi has a patch that's failing to apply

Bug: 168927099
Test: m, TH
Change-Id: If0ce56b4bb6edcdcea31b88d85cefb72e7e62a4e

*system/tools/aidl updated to 2b44fc81905023b3375c14a0b4103a809552eccf
Update test for build system change

build/soong/rust is being changed to use
`--extern force:foo=path/to/foo.dylib` to decrease the number of patches
we need to carry on 3p crates by allowing `dylib`s to be built out of
`#![no_std]` code naturally. The flag check here needed to be changed to
accept that flag.

Bug: 168927099
Test: m blueprint_tests
Change-Id: Id271af6776f31dfed25c355e3ad7d7da9b2deb1d