hwbcc: srv: introduce HwBccOps
Introduces a new trait for implementers of a hwbcc service.
Initially defines, init, close, and get_bcc. More required functions
will be added in later changes. HwBccOps are called based
on incoming commands found in the message header of each request. The
return values are packaged into a HwBccResponse and serialized, so this
change also includes serialization logic for HwBccResponse.

Bug: 369146791
Test: trusty/vendor/google/aosp/scripts/build.py qemu-generic-arm64-test-debug \
        --test="boot-test:com.android.trusty.rust.hwbcc.test"

Change-Id: I4935dce4d9db2366d6bc5d2e7762ed64fc777d73