lk: aarch64: use asm volatile to guard xpaci instruction
Volatile is needed to prevent compiler optimisation unconditionally
executing the `xpaci` instruction followed by `csel` to use or discard
the result.

Particularly the guarding to check for PAC support in dump_memory_around_register() before using `xpaci` becomes ineffective with newer clang versions ("18.0.x") and so the exception handler can cause an exception and end up in an infinite loop.

Bug: 388667671
Test: trusty-ut-ctrl com.android.trusty.crashtest
Test: manual inspection of assembly
Change-Id: Idad291d38c9625ff42823650507f78203f7b407a